requirements.txt
# Основные зависимости для Telegram бота астролога
# Установка: pip install -r requirements.txt

# Telegram Bot Framework
aiogram==3.2.0

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
asyncpg==0.29.0

# Валидация данных
pydantic==2.5.0

# HTTP клиент
aiohttp==3.9.1
requests==2.31.0

# Утилиты
python-dotenv==1.0.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1

# Логирование
loguru==0.7.2

# Дата и время
python-dateutil==2.8.2

# JSON
orjson==3.9.10

# Дополнительные библиотеки для астрологии
requests-html==0.10.0
